+++ packages/CopyInput/src/CopyInput.js
@@ -8,11 +8,15 @@ import useI18n from "@paprika/l10n/lib/useI18n";
 import { extractChildrenProps } from "@paprika/helpers";
 import CopyIcon from "@paprika/icon/lib/Clipboard";
 import CopyInputInputPropsCollector from "./components/Input/Input";
+import CopyInputButtonPropsCollector from "./components/Button/Button";
+import CopyInputPopoverPropsCollector from "./components/Popover/Popover";
 import * as sc from "./CopyInput.styles";
 
 function CopyInput(props) {
-  const { children, isReadOnly, value, ...moreProps } = props;
+  const { children, isReadOnly, hasInputContainer, hasValueContainer, value, ...moreProps } = props;
   const extendedInputProps = extractChildrenProps(children, CopyInputInputPropsCollector);
+  const extendedButtonProps = extractChildrenProps(children, CopyInputButtonPropsCollector);
+  const extendedPopoverProps = extractChildrenProps(children, CopyInputPopoverPropsCollector);
   const I18n = useI18n();
   const inputRef = React.createRef();
   const buttonRef = React.createRef();
@@ -32,26 +36,36 @@ function CopyInput(props) {
 
   function handleButtonClick() {
     const textToCopy = inputRef.current.value;
-    const textbox = document.createElement("textarea");
-    const activator = document.activeElement;
-    document.body.appendChild(textbox);
-    textbox.value = textToCopy;
-    textbox.select();
-    activator.focus();
-    document.execCommand("copy");
-    document.body.removeChild(textbox);
-
+    if (navigator.clipboard) {
+      navigator.clipboard.writeText(textToCopy);
+    } else {
+      // Fall back browser support
+      const textbox = document.createElement("textarea");
+      const activator = document.activeElement;
+      document.body.appendChild(textbox);
+      textbox.value = textToCopy;
+      textbox.select();
+      activator.focus();
+      document.execCommand("copy");
+      document.body.removeChild(textbox);
+    }
     setIsClickedTooltipOpen(true);
     setIsHoverTooltipOpen(false);
   }
 
   return (
-    <sc.CopyInput data-pka-anchor="copy-input" {...moreProps}>
-      <Input ref={inputRef} defaultValue={value} isReadOnly={isReadOnly} {...extendedInputProps} />
+    <sc.CopyInput data-pka-anchor="copy-input" hasDefaultButtonBorder={!hasInputContainer} {...moreProps}>
+      {hasValueContainer ? <sc.Value>{value}</sc.Value> : null}
+      {!hasInputContainer ? (
+        <sc.HiddenInput ref={inputRef} defaultValue={value} />
+      ) : (
+        <Input isReadOnly={isReadOnly} ref={inputRef} defaultValue={value} {...extendedInputProps} />
+      )}
       <div ref={buttonRef} data-pka-anchor="copy-input.button">
         <Button.Icon
           a11yText={I18n.t("copyInput.hover_tooltip")}
           kind="primary"
+          {...extendedButtonProps}
           onClick={handleButtonClick}
           onMouseOver={() => setIsHoverTooltipOpen(true)}
           onMouseOut={() => setIsHoverTooltipOpen(false)}
@@ -67,6 +81,7 @@ function CopyInput(props) {
         isDark
         isEager
         isOpen={isHoverTooltipOpen && !isClickedTooltipOpen}
+        {...extendedPopoverProps}
       >
         <Popover.Content>
           <Popover.Card>{I18n.t("copyInput.hover_tooltip")}</Popover.Card>
@@ -78,6 +93,7 @@ function CopyInput(props) {
         getPositioningElement={() => buttonRef.current}
         isOpen={isClickedTooltipOpen}
         shouldKeepFocus
+        {...extendedPopoverProps}
       >
         <Popover.Content>
           <Popover.Card>{I18n.t("copyInput.clicked_tooltip")}</Popover.Card>
@@ -85,7 +101,7 @@ function CopyInput(props) {
         <Popover.Tip />
       </Popover>
       {isClickedTooltipOpen && (
-        <Toast kind={Toast.types.kind.VISUALLY_HIDDEN} aria-hidden>
+        <Toast kind={Toast.types.kind.VISUALLY_HIDDEN} aria-hidden isPolite>
           {I18n.t("copyInput.clicked_tooltip")}
         </Toast>
       )}
@@ -98,6 +114,10 @@ const propTypes = {
   children: PropTypes.node,
   /** Is the input read-only. */
   isReadOnly: PropTypes.bool,
+  /** If the value will be rendered in an Input component or hidden */
+  hasInputContainer: PropTypes.bool,
+  /** If a plain text version of the value will be rendered */
+  hasValueContainer: PropTypes.bool,
   /** Default value for the input */
   value: PropTypes.string,
 };
@@ -105,6 +125,8 @@ const propTypes = {
 const defaultProps = {
   children: null,
   isReadOnly: true,
+  hasInputContainer: true,
+  hasValueContainer: false,
   value: "",
 };
 
@@ -113,5 +135,7 @@ CopyInput.propTypes = propTypes;
 CopyInput.defaultProps = defaultProps;
 
 CopyInput.Input = CopyInputInputPropsCollector;
+CopyInput.Button = CopyInputButtonPropsCollector;
+CopyInput.Popover = CopyInputPopoverPropsCollector;
 
 export default CopyInput;