+++ packages/CopyInput/src/components/Button/Button.js
@@ -1,0 +1,8 @@
+import React from "react";
+
+// Shell component for enhancing DX (developer experience)
+export default function Button() {
+  return <></>;
+}
+
+Button.displayName = "CopyInput.Button";