+++ packages/Table/src/index.js
@@ -1,1 +1,2 @@
+export { TH } from "./components/Headers/Headers.styles";
 export { default } from "./Table";