+++ packages/Breadcrumbs/CHANGELOG.md
@@ -1,5 +1,73 @@
 # Changelog
 
+## 1.2.0
+
+### Minor Changes
+
+- b00b8ce: - Support using any custom component/element as Breadcrumbs.Link
+  - Remove the arrow icon
+
+### Patch Changes
+
+- Updated dependencies [ff8eae9]
+- Updated dependencies [ff8eae9]
+  - @paprika/l10n@1.1.13
+  - @paprika/stylers@1.0.1
+  - @paprika/helpers@2.1.3
+  - @paprika/icon@1.3.0
+  - @paprika/button@1.1.2
+  - @paprika/popover@1.2.4
+
+## 1.2.0-next.1
+
+### Patch Changes
+
+- Updated dependencies [ff8eae9]
+- Updated dependencies [ff8eae9]
+  - @paprika/l10n@1.1.13-next.0
+  - @paprika/stylers@1.0.1-next.0
+  - @paprika/helpers@2.1.3-next.0
+  - @paprika/icon@1.3.0-next.0
+  - @paprika/button@1.1.2-next.0
+  - @paprika/popover@1.2.4-next.0
+
+## 1.2.0-next.0
+
+### Minor Changes
+
+- b00b8ce: - Support using any custom component/element as Breadcrumbs.Link
+  - Remove the arrow icon
+
+## 1.1.19
+
+### Patch Changes
+
+- Updated dependencies [b20c1d4]
+  - @paprika/l10n@1.1.12
+
+## 1.1.19-next.0
+
+### Patch Changes
+
+- Updated dependencies [b20c1d4]
+  - @paprika/l10n@1.1.12-next.0
+
+## 1.1.18
+
+### Patch Changes
+
+- Updated dependencies [009352de9]
+  - @paprika/icon@1.2.0
+  - @paprika/button@1.1.1
+
+## 1.1.18-next.0
+
+### Patch Changes
+
+- Updated dependencies [009352d]
+  - @paprika/icon@1.2.0-next.0
+  - @paprika/button@1.1.1-next.0
+
 ## 1.1.17
 
 ### Patch Changes