+++ packages/CopyInput/CHANGELOG.md
@@ -1,5 +1,105 @@
 # Changelog
 
+## 1.3.2
+
+### Patch Changes
+
+- Updated dependencies [ff8eae9]
+- Updated dependencies [ff8eae9]
+  - @paprika/l10n@1.1.13
+  - @paprika/helpers@2.1.3
+  - @paprika/icon@1.3.0
+  - @paprika/button@1.1.2
+  - @paprika/input@4.0.6
+  - @paprika/popover@1.2.4
+  - @paprika/toast@1.1.8
+
+## 1.3.2-next.0
+
+### Patch Changes
+
+- Updated dependencies [ff8eae9]
+- Updated dependencies [ff8eae9]
+  - @paprika/l10n@1.1.13-next.0
+  - @paprika/helpers@2.1.3-next.0
+  - @paprika/icon@1.3.0-next.0
+  - @paprika/input@4.0.6-next.0
+  - @paprika/toast@1.1.8-next.0
+  - @paprika/button@1.1.2-next.0
+  - @paprika/popover@1.2.4-next.0
+
+## 1.3.1
+
+### Patch Changes
+
+- Updated dependencies [b20c1d4]
+  - @paprika/l10n@1.1.12
+  - @paprika/input@4.0.5
+  - @paprika/toast@1.1.7
+
+## 1.3.1-next.0
+
+### Patch Changes
+
+- Updated dependencies [b20c1d4]
+  - @paprika/l10n@1.1.12-next.0
+  - @paprika/input@4.0.5-next.0
+  - @paprika/toast@1.1.7-next.0
+
+## 1.3.0
+
+### Minor Changes
+
+- a8878b8: Addding popover shell and use of clipboard API
+
+## 1.3.0-next.0
+
+### Minor Changes
+
+- a8878b8: Addding popover shell and use of clipboard API
+
+## 1.2.1
+
+### Patch Changes
+
+- Updated dependencies [bacebc9]
+  - @paprika/toast@1.1.6
+
+## 1.2.1-next.0
+
+### Patch Changes
+
+- Updated dependencies [bacebc9]
+  - @paprika/toast@1.1.6-next.0
+
+## 1.2.0
+
+### Minor Changes
+
+- 2c57bd573: Updating ability to show/hide the value text and ability to use the Input component or not; Adding Button shell component.
+
+### Patch Changes
+
+- Updated dependencies [009352de9]
+  - @paprika/icon@1.2.0
+  - @paprika/button@1.1.1
+  - @paprika/input@4.0.4
+  - @paprika/toast@1.1.5
+
+## 1.2.0-next.0
+
+### Minor Changes
+
+- 2c57bd5: Updating ability to show/hide the value text and ability to use the Input component or not; Adding Button shell component.
+
+### Patch Changes
+
+- Updated dependencies [009352d]
+  - @paprika/icon@1.2.0-next.0
+  - @paprika/button@1.1.1-next.0
+  - @paprika/input@4.0.4-next.0
+  - @paprika/toast@1.1.5-next.0
+
 ## 1.1.14
 
 ### Patch Changes