+++ packages/DataField/CHANGELOG.md
@@ -1,5 +1,64 @@
 <!-- content -->
 
+## 1.0.20
+
+### Patch Changes
+
+- Updated dependencies [ff8eae9]
+- Updated dependencies [ff8eae9]
+  - @paprika/data-grid@1.0.16
+  - @paprika/l10n@1.1.13
+  - @paprika/icon@1.3.0
+  - @paprika/button@1.1.2
+  - @paprika/popover@1.2.4
+
+## 1.0.20-next.0
+
+### Patch Changes
+
+- Updated dependencies [ff8eae9]
+- Updated dependencies [ff8eae9]
+  - @paprika/data-grid@1.0.16-next.0
+  - @paprika/l10n@1.1.13-next.0
+  - @paprika/icon@1.3.0-next.0
+  - @paprika/button@1.1.2-next.0
+  - @paprika/popover@1.2.4-next.0
+
+## 1.0.19
+
+### Patch Changes
+
+- Updated dependencies [b20c1d4]
+  - @paprika/l10n@1.1.12
+  - @paprika/data-grid@1.0.15
+
+## 1.0.19-next.0
+
+### Patch Changes
+
+- Updated dependencies [b20c1d4]
+  - @paprika/l10n@1.1.12-next.0
+  - @paprika/data-grid@1.0.15-next.0
+
+## 1.0.18
+
+### Patch Changes
+
+- Updated dependencies [659ebe184]
+- Updated dependencies [009352de9]
+  - @paprika/data-grid@1.0.14
+  - @paprika/icon@1.2.0
+  - @paprika/button@1.1.1
+
+## 1.0.18-next.0
+
+### Patch Changes
+
+- Updated dependencies [009352d]
+  - @paprika/icon@1.2.0-next.0
+  - @paprika/button@1.1.1-next.0
+  - @paprika/data-grid@1.0.14-next.0
+
 ## 1.0.17
 
 ### Patch Changes