+++ packages/DataGrid/CHANGELOG.md
@@ -1,5 +1,82 @@
 # Changelog
 
+## 1.0.16
+
+### Patch Changes
+
+- ff8eae9: Enable webpack tree shaking ability.
+- Updated dependencies [ff8eae9]
+- Updated dependencies [ff8eae9]
+  - @paprika/l10n@1.1.13
+  - @paprika/stylers@1.0.1
+  - @paprika/helpers@2.1.3
+  - @paprika/icon@1.3.0
+  - @paprika/button@1.1.2
+  - @paprika/checkbox@1.0.12
+  - @paprika/panel@2.1.3
+  - @paprika/raw-button@1.0.5
+  - @paprika/spinner@1.0.8
+
+## 1.0.16-next.0
+
+### Patch Changes
+
+- ff8eae9: Enable webpack tree shaking ability.
+- Updated dependencies [ff8eae9]
+- Updated dependencies [ff8eae9]
+  - @paprika/l10n@1.1.13-next.0
+  - @paprika/stylers@1.0.1-next.0
+  - @paprika/helpers@2.1.3-next.0
+  - @paprika/icon@1.3.0-next.0
+  - @paprika/panel@2.1.3-next.0
+  - @paprika/button@1.1.2-next.0
+  - @paprika/checkbox@1.0.12-next.0
+  - @paprika/raw-button@1.0.5-next.0
+  - @paprika/spinner@1.0.8-next.0
+
+## 1.0.15
+
+### Patch Changes
+
+- Updated dependencies [b20c1d4]
+  - @paprika/l10n@1.1.12
+  - @paprika/panel@2.1.2
+
+## 1.0.15-next.0
+
+### Patch Changes
+
+- Updated dependencies [b20c1d4]
+  - @paprika/l10n@1.1.12-next.0
+  - @paprika/panel@2.1.2-next.0
+
+## 1.0.14
+
+### Patch Changes
+
+- 659ebe184: Minor vertical alignment tweak of cells (header and body)
+- Updated dependencies [009352de9]
+  - @paprika/icon@1.2.0
+  - @paprika/button@1.1.1
+  - @paprika/checkbox@1.0.11
+  - @paprika/panel@2.1.1
+
+## 1.0.14-next.1
+
+### Patch Changes
+
+- 659ebe1: Minor vertical alignment tweak of cells (header and body)
+
+## 1.0.14-next.0
+
+### Patch Changes
+
+- Updated dependencies [009352d]
+  - @paprika/icon@1.2.0-next.0
+  - @paprika/button@1.1.1-next.0
+  - @paprika/checkbox@1.0.11-next.0
+  - @paprika/panel@2.1.1-next.0
+
 ## 1.0.13
 
 ### Patch Changes