+++ packages/ProgressBar/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @paprika/progress-bar
 
+## 1.1.2
+
+### Patch Changes
+
+- Updated dependencies [ff8eae9]
+  - @paprika/stylers@1.0.1
+  - @paprika/heading@1.0.7
+
+## 1.1.2-next.0
+
+### Patch Changes
+
+- Updated dependencies [ff8eae9]
+  - @paprika/stylers@1.0.1-next.0
+  - @paprika/heading@1.0.7-next.0
+
 ## 1.1.1
 
 ### Patch Changes